Added key and scale as alternative parameters to midi and mono
