+++ desktop/core/library/_keys.js
@@ -2,8 +2,6 @@
 
 import Operator from '../operator.js'
 
-const OCTAVE = ['C', 'c', 'D', 'd', 'E', 'F', 'f', 'G', 'g', 'A', 'a', 'B']
-
 export default function OperatorKeys (orca, x, y, passive) {
   Operator.call(this, orca, x, y, '&', true)
 
@@ -16,7 +14,7 @@ export default function OperatorKeys (orca, x, y, passive) {
     if (!terminal.io.midi.key) { return '.' }
     const octave = Math.floor(terminal.io.midi.key / 12)
     const value = terminal.io.midi.key % 12
-    const note = OCTAVE[value]
+    const note = this.notes[value]
     const transposed = this.transpose(note, octave)
     return transposed && transposed.real ? transposed.real : '.'
   }
