+++ desktop/core/scales.js
@@ -1,0 +1,73 @@
+'use strict'
+
+export default [
+    [0, 2, 4, 5, 7, 9, 11], //major
+    [0, 1, 3, 5, 7, 8, 10], //minor
+    [0, 3, 5, 7, 10], //minor_pentatonic
+    [0, 2, 5, 7, 9], //major_pentatonic
+    [0, 2, 3, 4, 7, 9], //blues_major
+    [0, 3, 5, 6, 7, 10], //blues_minor
+    [0, 1, 3, 5, 6, 8, 10], //dorian
+    [0, 2, 3, 5, 7, 8, 10], //phrygian
+    [0, 2, 4, 5, 7, 9, 10], //lydian
+    [0, 2, 4, 6, 7, 9, 11], //mixolydian
+    [0, 1, 3, 5, 7, 8, 10], //aeolian
+    [0, 2, 3, 5, 7, 9, 10], //locrian
+    [0, 2, 4, 5, 7, 9], //hex_major6
+    [0, 3, 5, 7, 8, 10], //hex_dorian
+    [0, 2, 5, 7, 9, 10], //hex_phrygian
+    [0, 2, 4, 7, 9, 11], //hex_major7
+    [0, 1, 3, 5, 8, 10], //hex_sus
+    [0, 2, 3, 5, 7, 10], //hex_aeolian
+    [0, 2, 4, 5, 7, 9, 10, 11], //bebop_dominant
+    [0, 3, 5, 8, 10], //egyptian
+    [0, 3, 5, 8, 10], //shang
+    [0, 2, 5, 7, 10], //jiao
+    [0, 2, 4, 7, 9], //zhi
+    [0, 2, 4, 7, 9], //ritusen
+    [0, 2, 4, 6, 8, 10], //whole
+    [0, 2, 3, 5, 7, 8, 11], //harmonic_minor
+    [0, 2, 3, 5, 7, 9, 11], //melodic_minor_asc
+    [0, 2, 3, 6, 7, 8, 11], //hungarian_minor
+    [0, 2, 3, 5, 6, 8, 9, 11], //octatonic
+    [0, 2, 4, 6, 8, 10], //messiaen1
+    [0, 1, 3, 4, 6, 7, 9, 10], //messiaen2
+    [0, 2, 3, 4, 6, 7, 8, 10, 11], //messiaen3
+    [0, 1, 2, 5, 6, 7, 8, 11], //messiaen4
+    [0, 1, 5, 6, 7, 11], //messiaen5
+    [0, 3, 4, 7, 8, 11], //augmented
+    [0, 2, 4, 5, 6, 8, 10, 11], //messiaen6
+    /* Not accessible after z=36 */
+    [0, 1, 2, 3, 5, 6, 7, 8, 9, 11], //messiaen7
+    [0, 1, 3, 4, 6, 8, 10], //super_locrian
+    [0, 2, 3, 7, 8], //hirajoshi
+    [0, 2, 3, 7, 9], //kumoi
+    [0, 1, 3, 5, 7, 9, 11], //neapolitan_major
+    [0, 2, 4, 5, 7, 8, 10], //bartok
+    [0, 1, 4, 5, 7, 8, 11], //bhairav
+    [0, 2, 4, 5, 6, 8, 10], //locrian_major
+    [0, 1, 4, 5, 7, 9, 10], //ahirbhairav
+    [0, 1, 4, 6, 8, 10, 11], //enigmatic
+    [0, 1, 3, 5, 7, 8, 11], //neapolitan_minor
+    [0, 1, 3, 7, 8], //pelog
+    [0, 1, 4, 5, 8, 9], //augmented2
+    [0, 1, 4, 7, 9], //scriabin
+    [0, 2, 4, 5, 7, 8, 11], //harmonic_major
+    [0, 2, 3, 5, 7, 8, 10], //melodic_minor_desc
+    [0, 2, 3, 6, 7, 9, 10], //romanian_minor
+    [0, 2, 4, 5, 7, 8, 10], //hindu
+    [0, 1, 5, 6, 10], //iwato
+    [0, 2, 3, 5, 7, 9, 11], //melodic_minor
+    [0, 2, 3, 5, 6, 8, 9, 11], //diminished2
+    [0, 1, 4, 6, 7, 9, 11], //marva
+    [0, 2, 4, 5, 7, 8, 10], //melodic_major
+    [0, 4, 5, 7, 10], //indian
+    [0, 1, 4, 5, 7, 8, 10], //spanish
+    [0, 2, 4, 6, 11], //prometheus
+    [0, 1, 3, 4, 6, 7, 9, 10], //diminished
+    [0, 1, 3, 6, 7, 8, 11], //todi
+    [0, 2, 4, 6, 8, 10, 11], //leading_whole
+    [0, 1, 4, 6, 7, 8, 11], //purvi
+    [0, 4, 6, 7, 11], //chinese
+    [0, 2, 4, 6, 7, 8, 10], //lydian_minor
+]
