+++ examples/_midi_degrees.orca
@@ -2,10 +2,7 @@
 ......6U8.................
 .......*..................
 ..9Uc..rg.................
-...*_016F5................
-..........................
-..........................
-..........................
+...*:013..F7..............
 ..........................
 ..........................
 ..........................
